adding asset mapping
