+++ mantis/simulation/routers/scaler.py
@@ -1,14 +1,14 @@
 
 # see https://or.stackexchange.com/questions/11603/numerical-infeasibility-for-moving-numbers-along-some-specific-conversion-edges
 
-    def scale_in():
-        """
-        Scales in data to be used by simulation
-        """
-        pass
-    
-    def scale_out():
-        """
-        Scales outs
-        """
-        pass    
+def scale_in():
+    """
+    Scales in data to be used by simulation
+    """
+    pass
+
+def scale_out():
+    """
+    Scales outs
+    """
+    pass
